useMovie hook created, movie component now uses that hook to handle logic
